# src/OpenBanking/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "API sets created by the Australian Consumer Data Standards to meet the needs of the Consumer Data Right",
    "version" : "1",
    "title" : "Consumer Data Standards",
    "license" : {
      "name" : "MIT Licence",
      "url" : "https://opensource.org/licenses/MIT"
    }
  },
  "host" : "data.provider.com.au",
  "basePath" : "/cds-au/v1",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/banking/accounts/{accountId}/direct-debits" : {
      "get" : {
        "tags" : [ "Direct Debits" ],
        "summary" : "Get Direct Debits For Account",
        "description" : "Obtain direct debit authorisations for a specific account",
        "operationId" : "listDirectDebits",
        "parameters" : [ {
          "name" : "accountId",
          "in" : "path",
          "description" : "ID of the account to get direct debit authorisations for.  Must have previously been returned by one of the account list end points.",
          "required" : true,
          "type" : "string",
          "format" : "ASCIIString"
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Page of results to request (standard pagination)",
          "required" : false,
          "type" : "integer",
          "default" : 1
        }, {
          "name" : "page-size",
          "in" : "query",
          "description" : "Page size to request. Default is 25 (standard pagination)",
          "required" : false,
          "type" : "integer",
          "default" : 25
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/ResponseBankingDirectDebits"
            }
          }
        },
        "x-scopes" : [ "bank_detailed_accounts" ],
        "x-version" : "1"
      }
    },
    "/banking/accounts/direct-debits" : {
      "get" : {
        "tags" : [ "Direct Debits" ],
        "summary" : "Get Bulk Direct Debits",
        "description" : "Obtain direct debit authorisations for multiple, filtered accounts",
        "operationId" : "listDirectDebitsBulk",
        "parameters" : [ {
          "name" : "is-owned",
          "in" : "query",
          "description" : "Filters accounts based on whether they are owned by the authorised customer",
          "required" : false,
          "type" : "string",
          "default" : "ALL",
          "enum" : [ "OWNED", "NOT_OWNED", "ALL" ]
        }, {
          "name" : "product-category",
          "in" : "query",
          "description" : "Used to filter results on the productCategory field applicable to accounts. Any one of the valid values for this field can be supplied. If absent then all accounts returned.",
          "required" : false,
          "type" : "string",
          "enum" : [ "PERS_AT_CALL_DEPOSITS", "BUS_AT_CALL_DEPOSITS", "TERM_DEPOSITS", "RESIDENTIAL_MORTGAGES", "PERS_CRED_AND_CHRG_CARDS", "BUS_CRED_AND_CHRG_CARDS", "PERS_LOANS", "PERS_LEASING", "BUS_LEASING", "TRADE_FINANCE", "PERS_OVERDRAFT", "BUS_OVERDRAFT", "BUS_LOANS", "FOREIGN_CURR_AT_CALL_DEPOSITS", "FOREIGN_CURR_TERM_DEPOSITS", "FOREIGN_CURR_LOAN", "FOREIGN_CURRRENCT_OVERDRAFT", "TRAVEL_CARD" ]
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Page of results to request (standard pagination)",
          "required" : false,
          "type" : "integer",
          "default" : 1
        }, {
          "name" : "page-size",
          "in" : "query",
          "description" : "Page size to request. Default is 25 (standard pagination)",
          "required" : false,
          "type" : "integer",
          "default" : 25
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/ResponseBankingDirectDebits"
            }
          }
        },
        "x-scopes" : [ "bank_detailed_accounts" ],
        "x-version" : "1"
      },
      "post" : {
        "tags" : [ "Direct Debits" ],
        "summary" : "Get Direct Debits For Specific Accounts",
        "description" : "Obtain direct debit authorisations for a specified list of accounts",
        "operationId" : "listDirectDebitsSpecificAccounts",
        "parameters" : [ {
          "in" : "body",
          "name" : "accountIds",
          "description" : "The list of account IDs to obtain information for",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/RequestAccountIds"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/ResponseBankingDirectDebits"
            }
          },
          "422" : {
            "description" : "The request was well formed but was unable to be processed due to business logic specific to the request",
            "schema" : {
              "$ref" : "#/definitions/ErrorList"
            }
          }
        },
        "x-scopes" : [ "bank_detailed_accounts" ],
        "x-version" : "1"
      }
    },
    "/banking/payees" : {
      "get" : {
        "tags" : [ "Payees" ],
        "summary" : "Get Payees",
        "description" : "Obtain a list of pre-registered payees",
        "operationId" : "listPayees",
        "parameters" : [ {
          "name" : "type",
          "in" : "query",
          "description" : "Filter on the payee type field.  In addition to normal type field values, ALL can be specified to retrieve all payees.  If absent the assumed value is ALL",
          "required" : false,
          "type" : "string",
          "default" : "ALL",
          "enum" : [ "DOMESTIC", "INTERNATIONAL", "BILLER", "ALL" ]
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Page of results to request (standard pagination)",
          "required" : false,
          "type" : "integer",
          "default" : 1
        }, {
          "name" : "page-size",
          "in" : "query",
          "description" : "Page size to request. Default is 25 (standard pagination)",
          "required" : false,
          "type" : "integer",
          "default" : 25
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/ResponseBankingPayees"
            }
          }
        },
        "x-scopes" : [ "bank_basic_accounts" ],
        "x-version" : "1"
      }
    },
    "/banking/payees/{payeeId}" : {
      "get" : {
        "tags" : [ "Payees" ],
        "summary" : "Get Payee Detail",
        "description" : "Obtain detailed information on a single payee",
        "operationId" : "getPayeeDetail",
        "parameters" : [ {
          "name" : "payeeId",
          "in" : "path",
          "description" : "The ID used to locate the details of a particular payee",
          "required" : true,
          "type" : "string",
          "format" : "ASCIIString"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/ResponseBankingPayeeDetails"
            }
          }
        },
        "x-scopes" : [ "bank_payees" ],
        "x-version" : "1"
      }
    },
...
  },
  "definitions" : {
    "RequestAccountIds" : {
      "type" : "object",
      "required" : [ "data", "meta" ],
      "properties" : {
        "data" : {
          "$ref" : "#/definitions/RequestAccountIds_data"
        },
        "meta" : {
          "$ref" : "#/definitions/Meta"
        }
      },
      "example" : {
        "data" : {
          "accountIds" : [ "accountIds", "accountIds" ]
        },
        "meta" : { }
      }
    },
...
    "ResponseBankingPayees" : {
      "type" : "object",
      "required" : [ "data", "links", "meta" ],
      "properties" : {
        "data" : {
          "$ref" : "#/definitions/ResponseBankingPayees_data"
        },
        "links" : {
          "$ref" : "#/definitions/LinksPaginated"
        },
        "meta" : {
          "$ref" : "#/definitions/MetaPaginated"
        }
      },
      "example" : {
        "data" : {
          "payees" : [ {
            "nickname" : "nickname",
            "description" : "description",
            "payeeId" : "payeeId",
            "type" : "DOMESTIC",
            "creationDate" : "creationDate"
          }, {
            "nickname" : "nickname",
            "description" : "description",
            "payeeId" : "payeeId",
            "type" : "DOMESTIC",
            "creationDate" : "creationDate"
          } ]
        },
        "meta" : {
          "totalRecords" : 0,
          "totalPages" : 6
        },
        "links" : {
          "next" : "next",
          "last" : "last",
          "prev" : "prev",
          "self" : "self",
          "first" : "first"
        }
      }
    },
    "ResponseBankingPayeeDetails" : {
      "type" : "object",
      "required" : [ "data", "links", "meta" ],
      "properties" : {
        "data" : {
          "$ref" : "#/definitions/PayeeDetail"
        },
        "links" : {
          "$ref" : "#/definitions/Links"
        },
        "meta" : {
          "$ref" : "#/definitions/Meta"
        }
      },
      "example" : {
        "data" : "",
        "meta" : { },
        "links" : {
          "self" : "self"
        }
      }
    },
    "Payee" : {
      "type" : "object",
      "required" : [ "nickname", "payeeId", "type" ],
      "properties" : {
        "payeeId" : {
          "type" : "string",
          "format" : "ASCIIString",
          "description" : "ID of the payee adhering to the rules of ID permanence"
        },
        "nickname" : {
          "type" : "string",
          "description" : "The short display name of the payee as provided by the customer"
        },
        "description" : {
          "type" : "string",
          "description" : "A description of the payee provided by the customer"
        },
        "type" : {
          "type" : "string",
          "description" : "The type of payee. DOMESTIC means a registered payee for domestic payments including NPP. INTERNATIONAL means a registered payee for international payments. BILLER means a registered payee for BPAY",
          "enum" : [ "DOMESTIC", "INTERNATIONAL", "BILLER" ]
        },
        "creationDate" : {
          "type" : "string",
          "format" : "DateString",
          "description" : "The date the payee was created by the customer"
        }
      },
      "example" : {
        "nickname" : "nickname",
        "description" : "description",
        "payeeId" : "payeeId",
        "type" : "DOMESTIC",
        "creationDate" : "creationDate"
      }
    },
    "PayeeDetail" : {
      "allOf" : [ {
        "$ref" : "#/definitions/Payee"
      }, {
        "type" : "object",
        "required" : [ "payeeUType" ],
        "properties" : {
          "payeeUType" : {
            "type" : "string",
            "description" : "Type of object included that describes the payee in detail",
            "enum" : [ "domestic", "biller", "international" ]
          },
          "domestic" : {
            "$ref" : "#/definitions/DomesticPayee"
          },
          "biller" : {
            "$ref" : "#/definitions/BillerPayee"
          },
          "international" : {
            "$ref" : "#/definitions/InternationalPayee"
          }
        },
        "x-conditional" : [ "domestic", "biller", "international" ]
      } ]
    },
    "DomesticPayee" : {
      "type" : "object",
      "required" : [ "payeeAccountUType" ],
      "properties" : {
        "payeeAccountUType" : {
          "type" : "string",
          "description" : "Type of account object included. Valid values are: { payeeAccountUType - - account A standard Australian account defined by BSB/Account Number payId A PayID recognised by NPP",
          "enum" : [ "account", "card", "payId" ]
        },
        "account" : {
          "$ref" : "#/definitions/DomesticPayeeAccount"
        },
        "card" : {
          "$ref" : "#/definitions/DomesticPayeeCard"
        },
        "payId" : {
          "$ref" : "#/definitions/DomesticPayeePayId"
        }
      },
      "x-conditional" : [ "account", "card", "payId" ]
    },
...
    "DomesticPayeeCard" : {
      "type" : "object",
      "required" : [ "cardNumber" ],
      "properties" : {
        "cardNumber" : {
          "type" : "string",
          "format" : "MaskedPANString",
          "description" : "Name of the account to pay to"
        }
      }
    },
...
    "BillerPayee" : {
      "type" : "object",
      "required" : [ "billerCode", "billerName" ],
      "properties" : {
        "billerCode" : {
          "type" : "string",
          "description" : "BPay Biller Code of the Biller"
        },
        "crn" : {
          "type" : "string",
          "description" : "BPay CRN of the Biller. If the contents of the CRN match the format of a Credit Card PAN then it should be masked using the rules applicable for the MaskedPANString common type"
        },
        "billerName" : {
          "type" : "string",
          "description" : "Name of the Biller"
        }
      },
      "x-conditional" : [ "crn" ]
    },
    "InternationalPayee" : {
      "type" : "object",
      "required" : [ "bankDetails", "beneficiaryDetails" ],
      "properties" : {
        "beneficiaryDetails" : {
          "$ref" : "#/definitions/InternationalPayee_beneficiaryDetails"
        },
        "bankDetails" : {
          "$ref" : "#/definitions/InternationalPayee_bankDetails"
        }
      }
    },
    "ResponseBankingDirectDebits" : {
      "type" : "object",
      "required" : [ "data", "links", "meta" ],
      "properties" : {
        "data" : {
          "$ref" : "#/definitions/ResponseBankingDirectDebits_data"
        },
        "links" : {
          "$ref" : "#/definitions/LinksPaginated"
        },
        "meta" : {
          "$ref" : "#/definitions/MetaPaginated"
        }
      },
      "example" : {
        "data" : {
          "directDebitAuthorisations" : [ {
            "lastDebitAmount" : "lastDebitAmount",
            "accountId" : "accountId",
            "lastDebitDateTime" : "lastDebitDateTime",
            "authorisedEntity" : {
              "arbn" : "arbn",
              "name" : "name",
              "financialInstitution" : "financialInstitution",
              "abn" : "abn",
              "acn" : "acn"
            }
          }, {
            "lastDebitAmount" : "lastDebitAmount",
            "accountId" : "accountId",
            "lastDebitDateTime" : "lastDebitDateTime",
            "authorisedEntity" : {
              "arbn" : "arbn",
              "name" : "name",
              "financialInstitution" : "financialInstitution",
              "abn" : "abn",
              "acn" : "acn"
            }
          } ]
        },
        "meta" : {
          "totalRecords" : 0,
          "totalPages" : 6
        },
        "links" : {
          "next" : "next",
          "last" : "last",
          "prev" : "prev",
          "self" : "self",
          "first" : "first"
        }
      }
    },
    "DirectDebits" : {
      "type" : "object",
      "required" : [ "accountId" ],
      "properties" : {
        "accountId" : {
          "type" : "string",
          "format" : "ASCIIString",
          "description" : "A unique ID of the account adhering to the standards for ID permanence."
        },
        "authorisedEntity" : {
          "$ref" : "#/definitions/AuthorisedEntity"
        },
        "lastDebitDateTime" : {
          "type" : "string",
          "format" : "DateTimeString",
          "description" : "The date and time of the last debit executed under this authorisation"
        },
        "lastDebitAmount" : {
          "type" : "string",
          "format" : "AmountString",
          "description" : "The amount of the last debit executed under this authorisation"
        }
      },
      "example" : {
        "lastDebitAmount" : "lastDebitAmount",
        "accountId" : "accountId",
        "lastDebitDateTime" : "lastDebitDateTime",
        "authorisedEntity" : {
          "arbn" : "arbn",
          "name" : "name",
          "financialInstitution" : "financialInstitution",
          "abn" : "abn",
          "acn" : "acn"
        }
      }
    },
...
    "Links" : {
      "type" : "object",
      "properties" : {
        "self" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "Fully qualified link to this API call"
        }
      },
      "example" : {
        "self" : "self"
      }
    },
    "Meta" : {
      "type" : "object"
    },
    "LinksPaginated" : {
      "type" : "object",
      "properties" : {
        "self" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "Fully qualified link to this API call"
        },
        "first" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "URI to the first page of this set. Mandatory if this response is not the first page"
        },
        "prev" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "URI to the first previous page of this set. Mandatory if this response is not the first page"
        },
        "next" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "URI to the first next page of this set. Mandatory if this response is not the last page"
        },
        "last" : {
          "type" : "string",
          "format" : "URIString",
          "description" : "URI to the first last page of this set. Mandatory if this response is not the last page"
        }
      },
      "example" : {
        "next" : "next",
        "last" : "last",
        "prev" : "prev",
        "self" : "self",
        "first" : "first"
      },
      "x-conditional" : [ "prev", "next", "first", "last" ]
    },
    "MetaPaginated" : {
      "type" : "object",
      "properties" : {
        "totalRecords" : {
          "type" : "integer",
          "format" : "NaturalNumber",
          "description" : "The total number of records in the full set"
        },
        "totalPages" : {
          "type" : "integer",
          "format" : "NaturalNumber",
          "description" : "The total number of pages in the full set"
        }
      },
      "example" : {
        "totalRecords" : 0,
        "totalPages" : 6
      }
    },
    "ErrorList" : {
      "type" : "object",
      "required" : [ "errors" ],
      "properties" : {
        "errors" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ErrorList_errors"
          }
        }
      }
    },
...
    "RequestAccountIds_data" : {
      "required" : [ "accountIds" ],
      "properties" : {
        "accountIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "ASCIIString",
            "description" : "A valid account ID previously obtained from another end point"
          }
        }
      },
      "example" : {
        "accountIds" : [ "accountIds", "accountIds" ]
      }
    },
...
    "ResponseBankingPayees_data" : {
      "required" : [ "payees" ],
      "properties" : {
        "payees" : {
          "type" : "array",
          "description" : "The list of payees returned",
          "items" : {
            "$ref" : "#/definitions/Payee"
          }
        }
      },
      "example" : {
        "payees" : [ {
          "nickname" : "nickname",
          "description" : "description",
          "payeeId" : "payeeId",
          "type" : "DOMESTIC",
          "creationDate" : "creationDate"
        }, {
          "nickname" : "nickname",
          "description" : "description",
          "payeeId" : "payeeId",
          "type" : "DOMESTIC",
          "creationDate" : "creationDate"
        } ]
      }
    },
...
    "ResponseBankingDirectDebits_data" : {
      "required" : [ "directDebitAuthorisations" ],
      "properties" : {
        "directDebitAuthorisations" : {
          "type" : "array",
          "description" : "The list of authorisations returned",
          "items" : {
            "$ref" : "#/definitions/DirectDebits"
          }
        }
      },
      "example" : {
        "directDebitAuthorisations" : [ {
          "lastDebitAmount" : "lastDebitAmount",
          "accountId" : "accountId",
          "lastDebitDateTime" : "lastDebitDateTime",
          "authorisedEntity" : {
            "arbn" : "arbn",
            "name" : "name",
            "financialInstitution" : "financialInstitution",
            "abn" : "abn",
            "acn" : "acn"
          }
        }, {
          "lastDebitAmount" : "lastDebitAmount",
          "accountId" : "accountId",
          "lastDebitDateTime" : "lastDebitDateTime",
          "authorisedEntity" : {
            "arbn" : "arbn",
            "name" : "name",
            "financialInstitution" : "financialInstitution",
            "abn" : "abn",
            "acn" : "acn"
          }
        } ]
      }
    },
...
    "ErrorList_errors" : {
      "required" : [ "code", "detail", "title" ],
      "properties" : {
        "code" : {
          "type" : "string",
          "description" : "The code of the error"
        },
        "title" : {
          "type" : "string",
          "description" : "A displayable title of the error type"
        },
        "detail" : {
          "type" : "string",
          "description" : "Detail of the error"
        },
        "meta" : {
          "type" : "object",
          "description" : "Optional additional data for specific error types",
          "properties" : { }
        }
      }
    }
  },
...
}